requirements.txt
aiogram==3.0.0
python-dotenv==1.0.0
Pillow==10.1.0
opencv-python==4.8.1
numpy==1.24.3
ultralytics==8.0.196
scikit-learn==1.3.0